feat: support mixed mode & updates
This change add support for the mixed-mode demo by adding a new
option to the terraform provisioning in Equinix Metal so that you can
create hosts to act as bare metal hosts. These baremetal hosts will have
the BYOH agent downloaded to them in preparation of the demo.

I have also also update the docs/terraform/scripts with the following
changes:
- Bumped the required terraform version (required by the next item)
- Bumped ubuntu from 20.04 to 22.04
- Changed the CIDR block for the private vlan from 192.168.1.x to
  192.168.10.x as 192.168.1.x can clash with a lot of home networks.

Signed-off-by: Richard Case <richard.case@outlook.com>